refactor : 아톰 개별 스토리스 폴더생성
