+++ src/components/UI/atoms/badges/stories/JobBadge.stories.js
@@ -1,4 +1,4 @@
-import JobBadge from './JobBadge';
+import JobBadge from '../JobBadge';
 
 export default {
   title: 'atoms/뱃지',
