+++ src/components/UI/atoms/buttons/stories/CloseButton.stories.js
@@ -1,4 +1,4 @@
-import CloseButton from './CloseButton';
+import CloseButton from '../CloseButton';
 
 export default {
   title: 'atoms/버튼',
